requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlmodel==0.0.16
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
pandas==2.2.1
openpyxl==3.1.2
qrcode==7.4.2
Pillow==10.2.0
supabase==2.3.6
python-multipart==0.0.9
alembic==1.13.1
Jinja2==3.1.3
python-jose[cryptography]==3.3.0
httpx==0.27.0